TST: switch from 'caproto' to 'softioc' IOC
