+++ pyproject.toml
@@ -41,7 +41,8 @@ dev = [
   "pytest-asyncio",
   "pre-commit",
   "ruff",
-  "caproto !=1.2.0",
+  # "caproto !=1.2.0",
+  "softioc",
   "pyepics",
 ]
 docs = [
@@ -62,7 +63,7 @@ minversion = "6.0"
 # Run pytest with all our checkers, and don't spam us with massive tracebacks on error
 # Don't collect the interactive directory which is intended for manual execution
 addopts = """
-    --tb=native -vv --doctest-modules --doctest-glob="*.rst" --ignore docs
+    --tb=native -vv --doctest-modules --doctest-glob="*.rst" --ignore docs --ignore ioc
     """
 # https://iscinumpy.gitlab.io/post/bound-version-constraints/#watch-for-warnings
 filterwarnings = [
